requirements.txt
django>=4.2, <5.0
djangorestframework==3.15.1
pyyaml==6.0.1
requests==2.32.3
django-cors-headers==4.3.1